Update Analyzing-world-s-best-wine README.md
# Analyzing-world-s-best-wine
This project will analyze a dataset of wine reviews to identify the world's best wine. The project will use data engineering skills to clean, wrangle, and visualize the data, and then use machine learning to predict the quality of wine.


Project Name: Analyzing the world's best wine
Technologies: Python, SQL, Pandas, NumPy, Matplotlib, Seaborn
Skills: Data cleaning,  Data wrangling, Data visualization, Machine learning.

Objective:
The objective of this project is to analyze a dataset of wine reviews to identify the world's best wine. The project will use data engineering skills to clean, wrangle, and visualize the data, and then use machine learning to predict the quality of wine.
             	Identify and understand the relationship between the wine industry and climate change.
             	Investigate the carbon emissions associated with the fermentation process in wine production.
             	Find insights and patterns in the data that can contribute to mitigating climate change effects in the wine industry.
             	Explore alternative approaches, such as the adoption of new grape varieties, to address the challenges posed by climate change.
             	Provide actionable recommendations to stakeholders in the wine industry for sustainable practices and decision-making.


Approach:
      	    Data collection: The first step will be to collect a dataset of wine reviews. There are many publicly available datasets that can be used for this project.
      	    Data cleaning: The next step will be to clean the data. This will involve removing any errors or inconsistencies in the data.
      	    Data wrangling: The data will then be wrangled to prepare it for analysis. This will involve transforming the data into a format that is easy to work with.
         	    Data visualization: The data will then be visualized to explore the relationships between the different variables. This will help to identify any trends or 
                patterns in the data.
      	    Machine learning: Finally, machine learning will be used to predict the quality of wine. This will involve training a model on the data and then using the 
                model to predict the quality of new wines.

Expected Outcomes:
The expected outcomes of this project are:
      •	A better understanding of the factors that contribute to the quality of wine.
      •	A list of the world's best wines.
      •	A machine learning model that can predict the quality of wine.

Benefits:
The benefits of this project include:
       	    The ability to make more informed decisions about wine purchases.
       	    The ability to identify new wines that you might enjoy.
       	    The ability to learn more about the world of wine.

Timeline:
      	     The project is expected to take 4-6 weeks to complete.

Budget:
      	   The total budget for the project is $1,000.



      
Sure, here are the technologies, skills, technology, objective, and approach for your project on analyzing the world's best wine using data engineer skills:

Technologies:

Hadoop: A distributed file system that can be used to store large amounts of wine data.
Hive: A SQL-like language that can be used to query data stored in Hadoop.
Spark: A fast and scalable processing engine that can be used to analyze wine data.
Tableau: A visualization tool that can be used to create interactive dashboards of wine data.
Python: A programming language that is often used for data analysis.
Skills:

SQL: A language for querying and manipulating wine data.
Data warehousing: The process of storing and managing large amounts of wine data.
Data architecture: The design of wine data systems.
Coding: The ability to write code in a programming language.
Machine learning: The ability to build models that can learn from wine data and make predictions.
Communication: The ability to communicate the results of your analysis to others.